# OpenAPI.yaml
openapi: 3.0.0
info:
  title: SyncYomi Server API
  description: SyncYomi Server API
  contact:
    name: Kaiser
    email: kaiserbh@proton.me
    url: https://github.com/flurbudurbur/Shiori
  version: 1.0.0
servers:
  - url: http://localhost:8282/api
paths:
...
  /notification:
    get:
      summary: List notifications
      description: List notifications
      operationId: listNotifications
      tags:
        - Notifications
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        default:
          description: Unexpected error
    post:
      summary: Store notification
      description: Store notification
      operationId: storeNotification
      tags:
        - Notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notification'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        default:
          description: Unexpected error
  /notification/test:
    post:
      summary: Test notification
      description: Test notification
      operationId: testNotification
      tags:
        - Notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notification'
      responses:
        '204':
          description: No Content
        default:
          description: Unexpected error
  /notification/{notificationID}:
    put:
      summary: Update notification
      description: Update notification
      operationId: updateNotification
      tags:
        - Notifications
      parameters:
        - name: notificationID
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notification'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        default:
          description: Unexpected error
    delete:
      summary: Delete notification
      description: Delete notification
      operationId: deleteNotification
      tags:
        - Notifications
      parameters:
        - name: notificationID
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
        default:
          description: Unexpected error
...
components:
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum:
            - DISCORD
            - NOTIFIARR
            - IFTTT
            - JOIN
            - MATTERMOST
            - MATRIX
            - PUSH_BULLET
            - PUSHOVER
            - ROCKETCHAT
            - SLACK
            - TELEGRAM
        enabled:
          type: boolean
        events:
          type: array
          items:
            type: string
            enum:
              - SERVER_UPDATE_AVAILABLE
              - SYNC_STARTED
              - SYNC_SUCCESS
              - SYNC_FAILED
              - SYNC_ERROR
              - TEST
        token:
          type: string
        apiKey:
          type: string
        webhook:
          type: string
        title:
          type: string
        icon:
          type: string
        username:
          type: string
        host:
          type: string
        password:
          type: string
        channel:
          type: string
        rooms:
          type: string
        targets:
          type: string
        devices:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...