# internal/http/notification.go
package http

import (
	"context"
	"encoding/json"
	"net/http"
	"strconv"

	"github.com/flurbudurbur/Shiori/internal/domain"
	"github.com/go-chi/chi/v5"
)

type notificationService interface {
	Find(context.Context, domain.NotificationQueryParams) ([]domain.Notification, int, error)
	FindByID(ctx context.Context, id int) (*domain.Notification, error)
	Store(ctx context.Context, n domain.Notification) (*domain.Notification, error)
	Update(ctx context.Context, n domain.Notification) (*domain.Notification, error)
	Delete(ctx context.Context, id int) error
	Test(ctx context.Context, notification domain.Notification) error
}

type notificationHandler struct {
	encoder encoder
	service notificationService
}

func newNotificationHandler(encoder encoder, service notificationService) *notificationHandler {
	return &notificationHandler{
		encoder: encoder,
		service: service,
	}
}

func (h notificationHandler) Routes(r chi.Router) {
	r.Get("/", h.list)
	r.Post("/", h.store)
	r.Post("/test", h.test)
	r.Put("/{notificationID}", h.update)
	r.Delete("/{notificationID}", h.delete)
}

func (h notificationHandler) list(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	user, ok := ctx.Value("user").(*domain.User)
	if !ok || user == nil {
		http.Error(w, "Unauthorized: User not found in context", http.StatusUnauthorized)
		return
	}

	list, _, err := h.service.Find(ctx, domain.NotificationQueryParams{UserHashedUUID: &user.HashedUUID})
	if err != nil {
		h.encoder.StatusNotFound(ctx, w)
		return
	}

	h.encoder.StatusResponse(ctx, w, list, http.StatusOK)
}

func (h notificationHandler) store(w http.ResponseWriter, r *http.Request) {
	var (
		ctx  = r.Context()
		data domain.Notification
	)

	if err := json.NewDecoder(r.Body).Decode(&data); err != nil {
		h.encoder.StatusResponse(ctx, w, "Invalid request body", http.StatusBadRequest)
		return
	}

	user, ok := ctx.Value("user").(*domain.User)
	if !ok || user == nil {
		http.Error(w, "Unauthorized: User not found in context", http.StatusUnauthorized)
		return
	}
	data.UserHashedUUID = user.HashedUUID

	filter, err := h.service.Store(ctx, data)
	if err != nil {
		h.encoder.StatusResponse(ctx, w, "Failed to store notification", http.StatusInternalServerError)
		return
	}

	h.encoder.StatusResponse(ctx, w, filter, http.StatusCreated)
}

func (h notificationHandler) update(w http.ResponseWriter, r *http.Request) {
	var (
		ctx  = r.Context()
		data domain.Notification
	)

	if err := json.NewDecoder(r.Body).Decode(&data); err != nil {
		h.encoder.StatusResponse(ctx, w, "Invalid request body", http.StatusBadRequest)
		return
	}

	user, ok := ctx.Value("user").(*domain.User)
	if !ok || user == nil {
		http.Error(w, "Unauthorized: User not found in context", http.StatusUnauthorized)
		return
	}
	// Ensure the update is for the authenticated user's notification
	// This might require fetching the existing notification first to check UserHashedUUID
	// or relying on the service layer to handle this authorization.
	// For now, we'll assume the service layer handles it or the ID in the path implies ownership.
	// If the notification itself needs to be associated with the user for update,
	// ensure data.UserHashedUUID is set if it's a new association or checked if it's an existing one.
	// data.UserHashedUUID = user.HashedUUID // Potentially set this if the model requires it for update logic

	filter, err := h.service.Update(ctx, data)
	if err != nil {
		h.encoder.StatusResponse(ctx, w, "Failed to update notification", http.StatusInternalServerError)
		return
	}

	h.encoder.StatusResponse(ctx, w, filter, http.StatusOK)
}

func (h notificationHandler) delete(w http.ResponseWriter, r *http.Request) {
	var (
		ctx            = r.Context()
		notificationID = chi.URLParam(r, "notificationID")
	)

	id, _ := strconv.Atoi(notificationID)

	if err := h.service.Delete(ctx, id); err != nil {
		// Consider user context for deletion if notifications are user-specific
		// For now, assuming ID is sufficient and service layer handles authorization
		h.encoder.StatusResponse(ctx, w, "Failed to delete notification", http.StatusInternalServerError)
		return
	}

	h.encoder.StatusResponse(ctx, w, nil, http.StatusNoContent)
}

func (h notificationHandler) test(w http.ResponseWriter, r *http.Request) {
	var (
		ctx  = r.Context()
		data domain.Notification
	)

	if err := json.NewDecoder(r.Body).Decode(&data); err != nil {
		h.encoder.StatusResponse(ctx, w, "Invalid request body", http.StatusBadRequest)
		return
	}

	// If testing a notification requires user context (e.g., to use user-specific settings)
	user, ok := ctx.Value("user").(*domain.User)
	if !ok || user == nil {
		// If user context is strictly required for testing, return error
		// http.Error(w, "Unauthorized: User not found in context for test", http.StatusUnauthorized)
		// return
		// If test can proceed without user or with default, do nothing or log
	}
	// If the notification being tested should be associated with the user:
	if user != nil {
		data.UserHashedUUID = user.HashedUUID
	}

	err := h.service.Test(ctx, data)
	if err != nil {
		h.encoder.StatusResponse(ctx, w, "Failed to test notification: "+err.Error(), http.StatusInternalServerError)
		return
	}

	h.encoder.NoContent(w)
}